# Optimize and add new efficient visual tree extensions.