## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Improves on current `VisualExtensions` by eliminating LINQ usage and adding a few new functions.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
Only `IEnumerable` extension methods are available which leads to poor performance and many wasteful memory allocations.

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->

100% decrease in allocations when finding common visual ancestor (used by hit testing, etc.)

|                     Method |       Mean |     Error |    StdDev | Ratio |       Gen 0 | Gen 1 | Gen 2 |    Allocated |
|--------------------------- |-----------:|----------:|----------:|------:|------------:|------:|------:|-------------:|
| CommonAncestor_IEnumerable | 3,530.3 ms | 14.935 ms | 13.970 ms |  1.00 | 617000.0000 |     - |     - | 2582357104 B |
|  CommonAncestor_Optimized |   387.7 ms |  2.426 ms |  2.026 ms |  0.11 |           - |     - |     - |            - |

Results from benchmarks included in this PR:

|                       Method |         Mean |         Error |        StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |-------------:|--------------:|--------------:|--------:|------:|------:|----------:|
|      FindAncestorOfType_Linq |     695.0 us |     3.6113 us |     3.0156 us | 83.0078 |     - |     - |  348640 B |
| FindAncestorOfType_Optimized |     184.0 us |     0.8506 us |     0.7956 us |       - |     - |     - |         - |
|     FindCommonVisualAncestor | 392,541.6 us | 1,693.0954 us | 1,500.8850 us |       - |     - |     - |         - |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
After rethinking https://github.com/AvaloniaUI/Avalonia/pull/3247 I've decided to just modify current extension methods and drop custom visitors. It is simpler to just write that code inline and less confusing for future maintainers.

I did change few random places since I wanted to get a feel how we are using visual trees and which parts can be optimized.

## Checklist

- [x] Added unit tests (if possible)?
- [x] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation

Replaces https://github.com/AvaloniaUI/Avalonia/pull/3247.